Create readme.md
This personal portfolio is designed to serve as a comprehensive showcase of my skills, experience, and projects as a web developer and designer. The purpose of this portfolio is to present my professional journey in an interactive and visually appealing way. Here's a detailed look at what the portfolio offers:


Signed-off-by: Hament <147083817+Itsyourhament@users.noreply.github.com>